Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/scriptnonce-ignore-unsafeinline.sub-expected.txt

CONSOLE MESSAGE: line 61: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
This tests that a valid nonce disables inline JavaScript, even if 'unsafe-inline' is present.


Harness Error (TIMEOUT), message = null

TIMEOUT Expecting alerts: ["PASS (1/2)","PASS (2/2)", "violated-directive=script-src-elem"] Test timed out
